app/requirements.txt
paho-mqtt==1.6.1
redis==5.0.4
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
resend==2.19.0
